WIP: Rats 26.1 port — scaffolding, mechanical passes, registries, client architecture
- Build: ModDevGradle 2.0.141, MC 26.1.2 / NeoForge 26.1.2.22-beta, Java 25 toolchain,
  Gradle 9.2.1, Citadel 26.1 from mavenLocal (com.github.alexthe666:citadelneoforge).
- Mechanical passes applied: ResourceLocation->Identifier (161 files), RenderType factory
  moves (46), MobSpawnType->EntitySpawnReason, GameRules package move, NBT getXOr family,
  contains(k), ValueInput/ValueOutput save-load signatures, package-info cleanup (52).
- Registries: register(name, key -> ...) + withItemId/withBlockId ID wrapping (298 sites),
  EntityType.build(ResourceKey), new BlockEntityType<>, armor materials -> 26.1 plain
  ArmorMaterial records + repair tags + equipment assets, spawn eggs -> RatsSpawnEggItem
  + rats:spawn_egg_layer tint source.
- Client architecture: RatsEntityModelBridge + RatsClientKeys (render-state entity stash),
  RatsRenderType rewritten on RenderSetup/RenderPipelines, data-driven item model
  properties (rat_count/glint_type/soul) + 411 generated client item definitions.
- Resources: pack.mcmeta min/max_format, neoforge.mods.toml for 26.1, datagen package
  excluded except tag constant holders (stripped to constants).
- Known remaining: renderer/layer/particle conversions, entity API fixes (trades, boats,
  arrows, hurtServer), item/block/GUI fixes, worldgen records (StructureSpawnOverride,
  TrimMaterial, addCarver) — partially complete, agents were interrupted.

+++ src/main/java/com/github/alexthe666/rats/server/entity/ai/goal/RatFollowOwnerGoal.java
@@ -74,7 +74,7 @@ public class RatFollowOwnerGoal extends FollowOwnerGoal {
 	private boolean attemptTeleportEntity(Entity mount, int x, int y, int z) {
 		BlockPos pos = new BlockPos(x, y, z);
 		if (!this.canTeleportTo(pos)) return false;
-		mount.moveTo(x + 0.5D, y, z + 0.5D, mount.getYRot(), mount.getXRot());
+		mount.moveTo(x + 0.5D, (double) y, z + 0.5D, mount.getYRot(), mount.getXRot());
 		this.rat.getNavigation().stop();
 		return true;
 	}
